refactor(tui): extract pure label/count helpers into tui::text
AppState (the pure state layer) depended on render.rs (presentation) for
two pure helpers — local_row_label and comment_lines_count. Move them into
a new pure tui::text module that both layers import, so the dependency
direction is corrected (state no longer reaches into render). comment_lines
(returns ratatui Line) stays in render; the existing count-matches-builder
test still guards their agreement. Behaviour unchanged (441 tests).

Refs #164 (arch-3)
